refactor: Introduce builder for sequential response client
